ZIM files moved from cache instead of copied
If not using --keep-downloads, then ZIM files are moved instead of copied then removed
- reduces space usage
- faster if on same disk

+++ ideascube/serveradmin/catalog.py
@@ -170,9 +170,10 @@ class Handler:
         return getattr(settings, setting, default)
 
     @classmethod
-    def install(cls, package, download_path):
+    def install(cls, package, download_path, keep_downloads=False):
         os.makedirs(cls._install_dir, exist_ok=True)
-        package.install(download_path, cls._install_dir)
+        package.install(download_path, cls._install_dir,
+                        keep_downloads=keep_downloads)
 
     @classmethod
     def remove(cls, package):
@@ -277,7 +278,7 @@ class Package(metaclass=MetaRegistry):
         except ValueError:
             return self.size
 
-    def install(self, download_path, install_dir):
+    def install(self, download_path, install_dir, keep_downloads=False):
         raise NotImplementedError('Subclasses must implement this method')
 
     def remove(self, install_dir):
@@ -319,7 +320,7 @@ class BaseZim(Package, no_register=True):
 
 
 class ZippedZim(BaseZim, typename='zipped-zim'):
-    def install(self, download_path, install_dir):
+    def install(self, download_path, install_dir, keep_downloads=False):
         self.assert_is_zipfile(download_path)
 
         with zipfile.ZipFile(download_path, "r") as z:
@@ -359,11 +360,14 @@ class ZippedZim(BaseZim, typename='zipped-zim'):
 
 
 class Zim(BaseZim, typename='zim'):
-    def install(self, download_path, install_dir):
+    def install(self, download_path, install_dir, keep_downloads=False):
         zim_name = '{self.id}.zim'.format(self=self)
         dest_name = os.path.join(install_dir, zim_name)
 
-        shutil.copyfile(download_path, dest_name)
+        if not keep_downloads:
+            shutil.move(download_path, dest_name)
+        else:
+            shutil.copyfile(download_path, dest_name)
 
     def remove(self, install_dir):
         zimname = '{self.id}.zim'.format(self=self)
@@ -374,7 +378,7 @@ class SimpleZipPackage(Package, no_register=True):
     def get_root_dir(self, install_dir):
         return os.path.join(install_dir, self.id)
 
-    def install(self, download_path, install_dir):
+    def install(self, download_path, install_dir, keep_downloads=False):
         self.assert_is_zipfile(download_path)
 
         with zipfile.ZipFile(download_path, "r") as z:
@@ -419,8 +423,9 @@ class ZippedMedias(SimpleZipPackage, typename='zipped-medias'):
         Document.objects.filter(package_id=self.id).delete()
         super().remove(install_dir)
 
-    def install(self, download_path, install_dir):
-        super().install(download_path, install_dir)
+    def install(self, download_path, install_dir, keep_downloads=False):
+        super().install(download_path, install_dir,
+                        keep_downloads=keep_downloads)
         print('Adding medias to mediacenter database.')
         root = self.get_root_dir(install_dir)
         manifestfile = Path(root, 'manifest.yml')
@@ -702,7 +707,8 @@ class Catalog:
 
             try:
                 print('Installing {pkg}'.format(pkg=pkg))
-                handler.install(pkg, download_path)
+                handler.install(pkg, download_path,
+                                keep_downloads=keep_downloads)
                 used_handlers.add(handler)
 
             except Exception as e:
@@ -830,7 +836,8 @@ class Catalog:
 
             try:
                 print('Installing {upkg}'.format(upkg=upkg))
-                uhandler.install(upkg, download_path)
+                uhandler.install(upkg, download_path,
+                                 keep_downloads=keep_downloads)
                 used_handlers.add(uhandler)
 
             except Exception as e:
